move the writer functions to a separate module
